requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy
matplotlib
scikit-learn
cupy-cuda12x
tensorrt
pycuda
timm
pillow
git+https://github.com/VainF/Torch-Pruning.git
